move model dump parsing to seperate file
