+++ pygbmexpl/xgb/__init__.py
@@ -1,4 +1,5 @@
 import pygbmexpl.xgb.explainer
+import pygbmexpl.xgb.parser
 import pygbmexpl.xgb.predictions
 
-__all__ = ['explainer', 'predictions']
+__all__ = ['explainer', 'parser', 'predictions']
